Commiting changes before switching branch
